WW-3714 Deprecate and migrate assorted marker interfaces
