+++ core/src/main/java/com/opensymphony/xwork2/ModelDriven.java
@@ -19,18 +19,8 @@
 package com.opensymphony.xwork2;
 
 /**
- * ModelDriven Actions provide a model object to be pushed onto the ValueStack
- * in addition to the Action itself, allowing a FormBean type approach like Struts.
- *
- * @author Jason Carreira
+ * @deprecated since 6.7.0, use {@link org.apache.struts2.ModelDriven} instead.
  */
-public interface ModelDriven<T> {
-
-    /**
-     * Gets the model to be pushed onto the ValueStack instead of the Action itself.
-     *
-     * @return the model
-     */
-    T getModel();
-
+@Deprecated
+public interface ModelDriven<T> extends org.apache.struts2.ModelDriven<T> {
 }
