+++ core/src/main/java/com/opensymphony/xwork2/Preparable.java
@@ -19,19 +19,8 @@
 package com.opensymphony.xwork2;
 
 /**
- * Preparable Actions will have their <code>prepare()</code> method called if the {@link com.opensymphony.xwork2.interceptor.PrepareInterceptor}
- * is applied to the ActionConfig.
- *
- * @author Jason Carreira
- * @see com.opensymphony.xwork2.interceptor.PrepareInterceptor
+ * @deprecated since 6.7.0, use {@link org.apache.struts2.Preparable} instead.
  */
-public interface Preparable {
-
-    /**
-     * This method is called to allow the action to prepare itself.
-     *
-     * @throws Exception thrown if a system level exception occurs.
-     */
-    void prepare() throws Exception;
-    
+@Deprecated
+public interface Preparable extends org.apache.struts2.Preparable {
 }
