+++ core/src/main/java/org/apache/struts2/Preparable.java
@@ -16,11 +16,22 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.opensymphony.xwork2.interceptor;
+package org.apache.struts2;
 
 /**
- * @deprecated since 6.7.0, use {@link org.apache.struts2.interceptor.ValidationErrorAware} instead.
+ * Preparable Actions will have their <code>prepare()</code> method called if the {@link com.opensymphony.xwork2.interceptor.PrepareInterceptor}
+ * is applied to the ActionConfig.
+ *
+ * @author Jason Carreira
+ * @see com.opensymphony.xwork2.interceptor.PrepareInterceptor
  */
-@Deprecated
-public interface ValidationErrorAware extends org.apache.struts2.interceptor.ValidationErrorAware {
+public interface Preparable {
+
+    /**
+     * This method is called to allow the action to prepare itself.
+     *
+     * @throws Exception thrown if a system level exception occurs.
+     */
+    void prepare() throws Exception;
+
 }
