+++ core/src/main/java/org/apache/struts2/Validateable.java
@@ -16,11 +16,20 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.opensymphony.xwork2.interceptor;
+package org.apache.struts2;
 
 /**
- * @deprecated since 6.7.0, use {@link org.apache.struts2.interceptor.ValidationErrorAware} instead.
+ * Provides an interface in which a call for a validation check can be done.
+ *
+ * @author Jason Carreira
+ * @see com.opensymphony.xwork2.ActionSupport
+ * @see com.opensymphony.xwork2.interceptor.DefaultWorkflowInterceptor
  */
-@Deprecated
-public interface ValidationErrorAware extends org.apache.struts2.interceptor.ValidationErrorAware {
+public interface Validateable {
+
+    /**
+     * Performs validation.
+     */
+    void validate();
+
 }
